+++ src/FluxFlow.Components.FileSystem/Contracts/FileWriteMode.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Components.FileSystem.Contracts;
+
+public enum FileWriteMode
+{
+    Overwrite = 0,
+    Append = 1,
+    CreateNew = 2
+}