+++ src/FluxFlow.Components.FileSystem/FileSystemComponentTypes.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Engine.Definitions;
+
+namespace FluxFlow.Components.FileSystem;
+
+public static class FileSystemComponentTypes
+{
+    public static readonly NodeType FileWrite = new("file.write");
+}